Implement throwable warping.
There are still various bugs but probably not related to throwing.

For the most part this is quite straight forward, see the Warping
section in throwing.md. The tricky part is that we need to restore the
ignored collision set, but the standard ThrowableItemMessage does not
contain such information. For this, I added a new EntityAuxMessage
mechanism where we can have any entity emit auxilliary messages. It goes
by the same code path as other server messages except it's not actually
sent to the client in the game (only to the controller), so there's no
need to hook up serialization in the game.

Also refactor the projects a bit to get rid of the "ai" shared project.
It is no longer necessary to share it; the only thing being shared is
the Thrift definition file, and the compiled output is not even shared.
The reason to get rid of it is to make the patch a single DLL so it can
be easily reloaded with BepInEx ScriptEngine.

Also, properly implemented OnDestroy for the patch plugin, so that it
can be properly reloaded, allowing the code to be updated without
reloading the game.

Also fix PhysicalAttachment handling; previously we were not properly
getting the attachment parent.