Ensure that batched messages are flushed before capturing frame.
There was bug for example when throwing, that on frame N the throwing
happens, but on frame N + 1 the thrown object is detached and starts
flying. This causes warping to the end of frame N to be problematic,
because frame N + 1 would not know that the object was being thrown.
The chef ends up still carrying the object.

The reason was because when sending messages from local server to local
client, messages smaller than 255 bytes are batched, and delayed until
the frame's MultiplayerController::LateUpdate(). (Other messages are
received synchronously by the local client; this is vice versa for
local client-to-server messages.) However, our plugin's LateUpdate() is
not necessarily run after MultiplayerController::LateUpdate(). So, in
this commit we make sure to flush the batched messages buffer before
capturing state, and also after warping (because if warping introduces
some (likely not critical) messages, we don't want those messages to
pollute the execution of the next frame. We want them processed
immediately).

Also add a framerate controller slider on the UI.

Also fix a NRE bug in attachment warping.