Minor fixes to UI and a warp bug fix.
There was a bug where after warping off from the cannon, throwing does
not work. This is because there's a subtle logic in the game that when
the Use button is down when entering or exiting from a session
interactable, the Use button is suppressed until it is released. That
means if we warp away from a frame where the use button is down, the use
button will not function again after it is released again. This
mechanism may be useful for real game but we don't rely on that for the
TAS, so for simplicity we just reset the suppression flag to false
whenever we warp.

Also add a doc about why a burger bug exists, for fun.