Fix a mixing container bug.
The bug happens after warping to a state where the mixer is in the
middle of mixing, when adding a new item to the mixer doesn't work. The
reason is because the controller code used the same progress for mixing
handler and cooking handler, but the mixer has both. And the mixer of
course does not allow addition of new ingredients when cooking has
already begun (such as in an oven).

This is fixed by splitting the progress into individual cases.

At this point there does not seem to be any more bugs in the currently
supported warping components (not sure about cannon).

Two bugs remaining to be fixed:
 - Save dialog sometimes doesn't work... this is bad.
 - ActionEditor doesn't allow putting in a decimal place when using
   @bind:event="oninput"