Implement washing station warping.
This is mostly pretty straight forward after the previous refactoring on
interactables.

Also fix a bug where the level won't end. This was due to overpatching
IsSynchronising. (I knew this would be an issue somewhere!)

Also fix a bug from the original game where DestroyEntitiesMessage was
never sending a sublist of IDs because the Initialise function was
implemented wrong. This was critical for TAS.