Fix the drink dispenser bug, also some CSS improvements.
The drink dispenser bug happened because the PlacementItemSwitcher
shared the same message type as the PickupItemSwitcher, so because the
drink had both of them but I only warped one of them, the
PlacementItemSwitcher made the controller overwrite the index. This is
fixed by warping both at the same time. Unfortunately the controller
code doesn't carefully handle two components that share the same message
type on the same entity, but luckily this particular case should be OK
because in all cases the two kinds of switchers should have equal
indexes.