Make the game deterministic! (#1)
Main issues:

Need to make sure synchroniseList is called on every frame, otherwise it's dependent on frame alignment.
Need to skip all ServerWorldObjectSynchroniser logic because that is dependent on frame alignment.
Need to also skip MeshLerper. It does some unnecessary transform lerping on every frame.
Added a comparison framework to check that two save files are equal, to test determinism.

In my testing I also overwrote byte 0x3bbd51 to 0x01 in the UnityPlayer.dll so that the physics jobs queue is always called with only 1 thread. I'm not sure if that's significant.